+++ table/src/main/java/tech/ydb/table/query/DataQueryResult.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import tech.ydb.proto.ValueProtos;
 import tech.ydb.proto.table.YdbTable;
+import tech.ydb.table.query.stats.QueryStats;
 import tech.ydb.table.result.ResultSetReader;
 import tech.ydb.table.result.impl.ProtoValueReaders;
-
 
 /**
  * @author Sergey Polovko
@@ -15,10 +15,12 @@ public class DataQueryResult {
 
     private final String txId;
     private final List<ValueProtos.ResultSet> resultSets;
+    private final QueryStats queryStats;
 
     public DataQueryResult(YdbTable.ExecuteQueryResult result) {
         this.txId = result.getTxMeta().getId();
         this.resultSets = result.getResultSetsList();
+        queryStats = result.hasQueryStats() ? new QueryStats(result.getQueryStats()) : null;
     }
 
     public String getTxId() {
@@ -43,5 +45,13 @@ public class DataQueryResult {
 
     public boolean isEmpty() {
         return txId.isEmpty() && resultSets.isEmpty();
+    }
+
+    public QueryStats getQueryStats() {
+        return queryStats;
+    }
+
+    public boolean hasQueryStats() {
+        return queryStats != null;
     }
 }